requirements.txt
matplotlib
pandas
requests
tqdm
torch
numpy
pyarrow
onnx
thop
tensorrboard
python-dotenv
aiohttp
